Fixed temp file issue in System Tests
Temp file handling channel id wasn't properly deleted after execution and test wasn't starting when file wasn't present

+++ tests/util/systemTestsManager.ts
@@ -68,8 +68,14 @@ export class SystemTestsManager
 
 	public clear(): void
 	{
-		fs.unlinkSync(path.join(this.DIR, 'replacementBotChannelId'));
-		fs.rmdirSync(this.DIR);
+		if(fs.existsSync(path.join(this.DIR, 'replacementBotChannelId')))
+		{
+			fs.unlinkSync(path.join(this.DIR, 'replacementBotChannelId'));
+		}
+		if(fs.existsSync(this.DIR))
+		{
+			fs.rmdirSync(this.DIR);
+		}
 	}
 
 	private createTestChannel(): Promise<TextChannel>
